Restructure: Split legacy buttons.js into modular structure with centralized functions
- Created new buttons.js as the main entry point with global event handlers
- Created functions.js for shared utilities (e.g. help button replacement, remove button)
- Removed logic from old buttons.js and relocated it into dedicated modules
- Updated footer.html
