+++ footer.html
@@ -43,14 +43,15 @@
 <!-- JavaScript-Dateien -->
 <!-- Language switch -->
 <script src="js/language.js"></script>
-<!-- Funktionen der Buttons -->
-<script src="js/buttons.js"></script>
 <!-- Thesauri laden -->
 <script src="js/thesauri.js"></script>
 <!-- Validierungen laden -->
 <script type="module" src="js/validation.js"></script>
 <!-- autocomplete laden -->
 <script src="js/autocomplete.js"></script>
+<!-- Eventhandlers einbinden -->
+<script type="module" src="js/eventhandlers/buttons.js"></script>
+<script type="module" src="js/eventhandlers/functions.js"></script>
 <!-- Roles laden -->
 <script src="js/roles.js"></script>
 <!-- Load Originating Laboratories -->
