+++ js/eventhandlers/functions.js
@@ -1,0 +1,42 @@
+/**
+ * @file functions.js
+ * @description Shared utility functions used by multiple form group modules
+ * such as cloning rows, managing layout consistency, and updating overlay labels.
+ * 
+ * @module functions
+ */
+
+/**
+
+* Replaces help buttons in cloned rows with invisible placeholders.
+* This helps maintain the structure and prevents changes in field sizes.
+*
+* @param {jQuery} row - The cloned row from which to replace help buttons.
+* @param {string} [roundCornersClass="input-right-with-round-corners"] - The CSS class for rounded corners.
+*/
+export function replaceHelpButtonInClonedRows(row, roundCornersClass = "input-right-with-round-corners") {
+  row.find("span.input-group-text:has(i.bi-question-circle-fill)").each(function () {
+    $(this).replaceWith('<div class="input-group-text" style="visibility: hidden; width: 42px; height: 38px;"></div>');
+  });
+
+  row.find(".input-with-help")
+    .removeClass("input-right-no-round-corners")
+    .addClass(roundCornersClass);
+}
+
+/**
+* Creates the Remove button element.
+* @returns {jQuery} A jQuery object representing the Remove button.
+*/
+export function createRemoveButton() {
+  return $('<button type="button" class="btn btn-danger removeButton" style="width: 36px;">-</button>');
+}
+
+/**
+* Updates the labels on the map overlays to match the current row numbering.
+ */
+export function updateOverlayLabels() {
+  if (typeof window.updateOverlayLabels === 'function') {
+    window.updateOverlayLabels();
+  }
+}
